fix up imports and such
